Fixed preview resizing not being alpha-corrected
